+++ cupy/_core/_fusion_op.py
@@ -2,6 +2,7 @@ import string
 
 import numpy
 
+from cupy._core import _codeblock
 from cupy._core._fusion_variable import _TraceVariable
 from cupy._core._fusion_variable import _TraceArray
 from cupy._core._fusion_variable import _VariableSet
@@ -9,7 +10,6 @@ from cupy._core import _fusion_thread_local
 from cupy._core import _kernel
 from cupy._core import _reduction
 from cupy._core._scalar import get_typename
-from cupyx.jit import _codeblock
 
 
 class _UfuncRoutine:
